+++ require.php
@@ -1,14 +1,17 @@
 <?php
 require_once('src/Base/Base.php');
 require_once('src/Base/AncestorTrait.php');
-require_once('src/Base/BaseClass.php');
 
-require_once('src/ObjectTools/ObjectTools.php');
-require_once('src/FilesTools/FilesTools.php');
+require_once('src/Tools/Objects.php');
+require_once('src/Tools/Files.php');
+require_once('src/Tools/Cache.php');
+
 require_once('src/ObjectCollection/ObjectCollection.php');
 require_once('src/Response/Response.php');
 require_once('src/Storage/Storage.php');
 require_once('src/Storage/DB/Storage.php');
 
 require_once('src/Snippet/Snippet.php');
+
+require_once('src/backwardCompatibility.php');
 ?>