+++ src/backwardCompatibility.php
@@ -1,7 +1,9 @@
 <?php
 namespace DDTools;
 
-//Backward compatibility
 abstract class BaseClass extends \DDTools\Base\Base {
 	use \DDTools\Base\AncestorTrait;
 }
+
+class ObjectTools extends \DDTools\Tools\Objects {}
+class FilesTools extends \DDTools\Tools\Files {}